refactor: split onboard flow types
